+++ process_fits.py
@@ -92,7 +92,7 @@ for idx, image in enumerate(images):
         header_out['PROC'] = 'True'
         header_out['COMMENT'] = 'Processed %s on %s' % (get_Version(), strftime("%Y-%m-%dT%H-%M-%S"))
 
-        hdu = fits.PrimaryHDU(images[idx][0], header=header_out)
+        hdu = fits.PrimaryHDU(images[idx][0].astype(np.uint16), header=header_out)
         hdu.writeto(path_fits)
 
         # generate previews
